requirements.txt
tensorflow-gpu==2.4.1
matplotlib==3.7.0
numpy~=1.23.5
gym~=0.26.1
keras~=2.10.0
